+++ components/header.php
@@ -1,9 +1,13 @@
-    <div class="d-flex flex-column flex-md-row align-items-center p-3 px-md-4 mb-3 bg-white border-bottom shadow-sm">
-        <h5 class="d-none my-0 mr-md-auto font-weight-normal">Flash Chord</h5><img src="/images/flash-chord-logo.png" alt="Flash Chord" title="Flash Chord" class="my-0 mr-md-auto" />
-        <nav class="my-2 my-md-0 mr-md-3">
-            <a class="p-2 text-dark" href="/">Flash Chord</a>
-            <a class="p-2 text-dark" href="/donate">Donate</a>
-            <a class="p-2 text-dark" href="/faq">FAQ</a>
-            <a class="p-2 text-dark" href="/about">About</a>
-        </nav>
+    <div class="row p-3 px-md-4 mb-3 border-bottom shadow-sm align-items-center">
+        <div class="col text-start">
+            <h5 class="d-none my-0 mr-md-auto font-weight-normal">Flash Chord</h5><img src="/images/flash-chord-logo.png" alt="Flash Chord" title="Flash Chord" class="my-0 mr-md-auto" />
+        </div>
+        <div class="col text-end">
+            <nav class="my-2 my-md-0 mr-md-3">
+                <a class="p-2 text-dark" href="/">Flash Chord</a>
+                <a class="p-2 text-dark" href="/donate">Donate</a>
+                <a class="p-2 text-dark" href="/faq">FAQ</a>
+                <a class="p-2 text-dark" href="/about">About</a>
+            </nav>
+        </div>
     </div>